# backend/src/api/bridge-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Multi-Robot Control System - Device Bridge",
    "description": "Python bridge server for device communication",
    "version": "1.0.0"
  },
  "paths": {
...
    "/devices/{device_id}/grbl-settings": {
      "get": {
        "summary": "Get Grbl Settings",
        "description": "GRBL beállítások lekérdezése ($$).",
        "operationId": "get_grbl_settings_devices__device_id__grbl_settings_get",
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Device Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Set Grbl Setting",
        "description": "GRBL beállítás módosítása ($N=value).",
        "operationId": "set_grbl_setting_devices__device_id__grbl_settings_post",
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Device Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GrblSettingRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/devices/{device_id}/grbl-settings/batch": {
      "post": {
        "summary": "Set Grbl Settings Batch",
        "description": "Több GRBL beállítás módosítása egyszerre.",
        "operationId": "set_grbl_settings_batch_devices__device_id__grbl_settings_batch_post",
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Device Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GrblSettingsBatchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "GrblSettingRequest": {
        "properties": {
          "setting": {
            "type": "integer",
            "title": "Setting"
          },
          "value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              }
            ],
            "title": "Value"
          }
        },
        "type": "object",
        "required": [
          "setting",
          "value"
        ],
        "title": "GrblSettingRequest",
        "description": "Egyetlen GRBL beállítás módosítása ($N=value)."
      },
      "GrblSettingsBatchRequest": {
        "properties": {
          "settings": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "string"
                }
              ]
            },
            "type": "object",
            "title": "Settings"
          }
        },
        "type": "object",
        "required": [
          "settings"
        ],
        "title": "GrblSettingsBatchRequest",
        "description": "Több GRBL beállítás egyszerre.\n\nA networking paraméterek (pl. $71 hostname, $73-$76 SSID/jelszó) string\nértékűek a grblHAL-ban; ezért a value típus `Union[float, str]`."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}